requirements.txt
discord.py>=2.3.0
PyYAML>=6.0
